split Piece classes into their own module, same with unit tests.
